+++ app/php/class-setup.php
@@ -50,27 +50,29 @@ class Setup {
 
 		// add_action( 'wp_ajax_wp-live-debug-accept-risk', array( '\\WP_Live_Debug\\Page', 'accept_risk' ) );
 
-		// log related actions.
-		// add_action( 'wp_ajax_wp-live-debug-read-log', array( '\\WP_Live_Debug\\Debug_Log', 'read_debug_log' ) );
-		// add_action( 'wp_ajax_wp-live-debug-select-log', array( '\\WP_Live_Debug\\Debug_Log', 'select_log_file' ) );
-		// add_action( 'wp_ajax_wp-live-debug-clear-debug-log', array( '\\WP_Live_Debug\\Debug_Log', 'clear_debug_log' ) );
-		// add_action( 'wp_ajax_wp-live-debug-delete-debug-log', array( '\\WP_Live_Debug\\Debug_Log', 'delete_debug_log' ) );
-		// add_action( 'wp_ajax_wp-live-debug-refresh-debug-log', array( '\\WP_Live_Debug\\Debug_Log', 'refresh_debug_log' ) );
+		// Log related actions.
+		// add_action( 'wp_ajax_wp-live-debug-read-log', array( '\\WP_Live_Debug\\Log', 'read_debug_log' ) );
+		// add_action( 'wp_ajax_wp-live-debug-select-log', array( '\\WP_Live_Debug\\Log', 'select_log_file' ) );
+		// add_action( 'wp_ajax_wp-live-debug-clear-debug-log', array( '\\WP_Live_Debug\\Log', 'clear_debug_log' ) );
+		// add_action( 'wp_ajax_wp-live-debug-delete-debug-log', array( '\\WP_Live_Debug\\Log', 'delete_debug_log' ) );
+		// add_action( 'wp_ajax_wp-live-debug-refresh-debug-log', array( '\\WP_Live_Debug\\Log', 'refresh_debug_log' ) );
 
 		// Backup actions.
-		add_action( 'wp_ajax_wp-live-debug-create-backup', array( '\\WP_Live_Debug\\WP_Config', 'create_wp_config_backup' ) );
-		add_action( 'wp_ajax_wp-live-debug-restore-backup', array( '\\WP_Live_Debug\\WP_Config', 'restore_wp_config_backup' ) );
-		add_action( 'wp_ajax_wp-live-debug-check-auto-backup', array( '\\WP_Live_Debug\\WP_Config', 'check_wp_config_original_backup' ) );
-		//add_action( 'admin_init', array( '\\WP_Live_Debug\\WP_Config', 'download_config_backup' ) );
+		add_action( 'wp_ajax_wp-live-debug-create-backup', array( '\\WP_Live_Debug\\Config', 'create_wp_config_backup' ) );
+		add_action( 'wp_ajax_wp-live-debug-restore-backup', array( '\\WP_Live_Debug\\Config', 'restore_wp_config_backup' ) );
+		add_action( 'wp_ajax_wp-live-debug-check-auto-backup', array( '\\WP_Live_Debug\\Config', 'check_wp_config_original_backup' ) );
+		//add_action( 'admin_init', array( '\\WP_Live_Debug\\Config', 'download_config_backup' ) );
+
+		// Constant actions.
+		add_action( 'wp_ajax_wp-live-debug-is-constant-true', array( '\\WP_Live_Debug\\Constants', 'is_constant_true' ) );
 
 		// wp-config related actions.
-		add_action( 'wp_ajax_wp-live-debug-is-constant-true', array( '\\WP_Live_Debug\\WP_Config', 'is_constant_true' ) );
-		// add_action( 'wp_ajax_wp-live-debug-enable', array( '\\WP_Live_Debug\\WP_Config', 'enable_wp_debug' ) );
-		// add_action( 'wp_ajax_wp-live-debug-disable', array( '\\WP_Live_Debug\\WP_Config', 'disable_wp_debug' ) );
-		// add_action( 'wp_ajax_wp-live-debug-enable-script-debug', array( '\\WP_Live_Debug\\WP_Config', 'enable_script_debug' ) );
-		// add_action( 'wp_ajax_wp-live-debug-disable-script-debug', array( '\\WP_Live_Debug\\WP_Config', 'disable_script_debug' ) );
-		// add_action( 'wp_ajax_wp-live-debug-enable-savequeries', array( '\\WP_Live_Debug\\WP_Config', 'enable_savequeries' ) );
-		// add_action( 'wp_ajax_wp-live-debug-disable-savequeries', array( '\\WP_Live_Debug\\WP_Config', 'disable_savequeries' ) );
+		// add_action( 'wp_ajax_wp-live-debug-enable', array( '\\WP_Live_Debug\\Config', 'enable_wp_debug' ) );
+		// add_action( 'wp_ajax_wp-live-debug-disable', array( '\\WP_Live_Debug\\Config', 'disable_wp_debug' ) );
+		// add_action( 'wp_ajax_wp-live-debug-enable-script-debug', array( '\\WP_Live_Debug\\Config', 'enable_script_debug' ) );
+		// add_action( 'wp_ajax_wp-live-debug-disable-script-debug', array( '\\WP_Live_Debug\\Config', 'disable_script_debug' ) );
+		// add_action( 'wp_ajax_wp-live-debug-enable-savequeries', array( '\\WP_Live_Debug\\Config', 'enable_savequeries' ) );
+		// add_action( 'wp_ajax_wp-live-debug-disable-savequeries', array( '\\WP_Live_Debug\\Config', 'disable_savequeries' ) );
 	}
 
 	/**
@@ -79,8 +81,8 @@ class Setup {
 	public static function activate() {
 		update_option( 'wp_live_debug_auto_refresh', 'disabled' );
 
-		Helper::create_debug_log();
-		Helper::get_first_backup();
+		Log::create_debug_log();
+		Log::get_first_backup();
 	}
 
 	/**
@@ -91,7 +93,7 @@ class Setup {
 		delete_option( 'wp_live_debug_log_file' );
 		delete_option( 'wp_live_debug_auto_refresh' );
 
-		Helper::clear_manual_backup();
+		Log::clear_manual_backup();
 	}
 
 	/**
